Bring back the shared llvm::Module ownership model.
Before this patch, the llvm::LazyCompilationLayer can 'steal' the llvm::Module*
and we cannot use it anymore for understanding the delta of the incremental
changes in cling. Another issue is that we cannot register the at_exit funcs.

This patch brings back the shared_ptr ownership which suits us better until we
find a way forward.